requirements.txt
opencv-python
numpy
PyAutoGUI
keyboard
pywin32
requests
Pillow
ultralytics
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv